# shop-api-golang/docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CreateUserRequest:
    properties:
      email:
        type: string
      name:
        type: string
    required:
    - email
    - name
    type: object
...
  dto.CursorUserResponse:
    properties:
      next_cursor:
        type: integer
      users:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
    required:
    - next_cursor
    - users
    type: object
...
  dto.DailyUserRegistrationResponse:
    properties:
      count:
        type: integer
      createdAt:
        type: string
    required:
    - count
    - createdAt
    type: object
...
  dto.ListUserByMostExpensiveProductResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.UserByMostExpensiveProduct'
        type: array
      total:
        type: integer
    required:
    - data
    - total
    type: object
  dto.ListUserResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.UserResponse'
        type: array
      total:
        type: integer
    required:
    - data
    - total
    type: object
  dto.ListUserWithTotalResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.UserWithTotalSpent'
        type: array
      total:
        type: integer
    required:
    - data
    - total
    type: object
...
  dto.UserByMostExpensiveProduct:
    properties:
      email:
        example: text@gmail.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: username
        type: string
    required:
    - email
    - id
    - name
    type: object
...
  dto.UserResponse:
    properties:
      email:
        example: text@gmail.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: username
        type: string
    required:
    - email
    - id
    - name
    type: object
  dto.UserWithTotalSpent:
    properties:
      email:
        example: text@gmail.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: username
        type: string
      totalSpent:
        example: 123.123
        type: number
    required:
    - email
    - id
    - name
    - totalSpent
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Educational SQL workshop backend API
  title: Shop API
  version: "1.0"
paths:
...
  /users:
    get:
      description: Returns all users from the database
      parameters:
      - description: Number of users to return (default 10)
        in: query
        name: limit
        type: integer
      - description: Number of users to skip (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListUserResponse'
      security:
      - BearerAuth: []
      summary: List all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Creates a new user and publishes UserCreated event to Kafka
      parameters:
      - description: User data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a user
      tags:
      - users
  /users/by-most-expensive-product:
    get:
      description: Returns users by the most expensive product
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListUserByMostExpensiveProductResponse'
      security:
      - BearerAuth: []
      summary: Get users by the most expensive product
      tags:
      - users
...
  /users/cursor:
    get:
      description: Returns users with cursor-based pagination. Pass the last user
        ID from the previous response as cursor.
      parameters:
      - description: Number of users to return (default 10)
        in: query
        name: limit
        type: integer
      - description: Last users ID from previous page
        in: query
        name: cursor
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CursorUserResponse'
      security:
      - BearerAuth: []
      summary: List users (cursor-based)
      tags:
      - users
  /users/daily-registrations:
    get:
      description: Returns user registration count for today
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DailyUserRegistrationResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get daily user registration
      tags:
      - users
...
  /users/search:
    get:
      description: Search users by field (partial match)
      parameters:
      - description: Field to search by
        enum:
        - email
        - name
        in: query
        name: field
        required: true
        type: string
      - description: Value to search for
        in: query
        name: value
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListUserResponse'
      security:
      - BearerAuth: []
      summary: Search user by field
      tags:
      - users
  /users/top-3-users:
    get:
      description: Returns top 3 users with the highest number of total spent
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ListUserWithTotalResponse'
      security:
      - BearerAuth: []
      summary: Get top 3 users by total spent
      tags:
      - users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"